+++ src/onegov.election_day/onegov/election_day/layout.py
@@ -538,3 +538,15 @@ class ManageDataSourceItemsLayout(ManageLayout):
         self.breadcrumbs.append(
             (_("Mappings"), request.link(self.model), ''),
         )
+
+
+class MailLayout(Layout):
+    """ A special layout for creating HTML E-Mails. """
+
+    @cached_property
+    def base(self):
+        return self.template_loader['mail_layout.pt']
+
+    @cached_property
+    def primary_color(self):
+        return self.app.theme_options.get('primary-color', '#fff')