+++ src/onegov.election_day/onegov/election_day/tests/test_views_manage.py
@@ -1,4 +1,5 @@
 from datetime import date
+from lxml.html import document_fromstring
 from onegov.election_day.collections import ArchivedResultCollection
 from onegov.election_day.tests import login
 from webtest import TestApp as Client
@@ -242,3 +243,60 @@ def test_view_manage_data_sources(election_day_app):
     client.get('/manage/sources').click('Löschen').form.submit()
     assert 'ds_proporz' not in client.get('/manage/sources')
     assert 'Noch keine Datenquellen' in client.get('/manage/sources')
+
+
+def test_reset_password(election_day_app):
+    client = Client(election_day_app)
+    client.get('/locale/de_CH').follow()
+
+    request_page = client.get('/auth/login').click('Passwort zurücksetzen')
+    assert 'Passwort zurücksetzen' in request_page
+
+    request_page.form['email'] = 'someone@example.org'
+    assert 'someone@example.org' in request_page.form.submit()
+    assert len(election_day_app.smtp.outbox) == 0
+
+    request_page.form['email'] = 'admin@example.org'
+    assert 'admin@example.org' in request_page.form.submit()
+    assert len(election_day_app.smtp.outbox) == 1
+
+    message = election_day_app.smtp.outbox[0]
+    message = message.get_payload(1).get_payload(decode=True)
+    message = message.decode('iso-8859-1')
+    link = list(document_fromstring(message).iterlinks())[0][2]
+    token = link.split('token=')[1]
+
+    reset_page = client.get(link)
+    assert token in reset_page.text
+
+    reset_page.form['email'] = 'someone_else@example.org'
+    reset_page.form['password'] = 'new_password'
+    reset_page = reset_page.form.submit()
+    assert "Ungültige Addresse oder abgelaufener Link" in reset_page
+    assert token in reset_page.text
+
+    reset_page.form['email'] = 'admin@example.org'
+    reset_page.form['password'] = '1234'
+    reset_page = reset_page.form.submit()
+    assert "Feld muss mindestens 8 Zeichen beinhalten" in reset_page
+    assert token in reset_page.text
+
+    reset_page.form['email'] = 'admin@example.org'
+    reset_page.form['password'] = 'new_password'
+    assert "Passwort geändert" in reset_page.form.submit()
+
+    reset_page.form['email'] = 'admin@example.org'
+    reset_page.form['password'] = 'new_password'
+    reset_page = reset_page.form.submit()
+    assert "Ungültige Addresse oder abgelaufener Link" in reset_page
+
+    login_page = client.get('/auth/login')
+    login_page.form['username'] = 'admin@example.org'
+    login_page.form['password'] = 'hunter2'
+    login_page = login_page.form.submit()
+    assert "Unbekannter Benutzername oder falsches Passwort" in login_page
+
+    login_page.form['username'] = 'admin@example.org'
+    login_page.form['password'] = 'new_password'
+    login_page = login_page.form.submit().follow()
+    assert "Sie sind angemeldet" in login_page